refactor: centralize theme settings mixin
