+++ posawesome/public/js/posapp/components/pos/Customer.vue
@@ -132,8 +132,10 @@
 <script>
 import UpdateCustomer from './UpdateCustomer.vue';
 import { getCustomerStorage, setCustomerStorage } from '../../../offline/index.js';
+import { themeSettingsMixin } from '../../mixins/themeSettings.js';
 
 export default {
+  mixins: [themeSettingsMixin],
   props: {
     pos_profile: Object
   },
@@ -156,10 +158,6 @@ export default {
   },
 
   computed: {
-    isDarkTheme() {
-      return this.$theme.current === 'dark';
-    },
-
     filteredCustomers() {
       const search = this.customerSearch.toLowerCase();
       let results = this.customers;
