+++ posawesome/public/js/posapp/components/pos/ItemsTable.vue
@@ -201,8 +201,10 @@
 </template>
 
 <script>
+import { themeSettingsMixin } from '../../mixins/themeSettings.js';
 export default {
   name: 'ItemsTable',
+  mixins: [themeSettingsMixin],
   props: {
     headers: Array,
     items: Array,
@@ -245,21 +247,7 @@ export default {
         ? { style: 'background-color:#121212;color:#fff' }
         : {};
     },
-    isDarkTheme() {
-      return this.$theme.current === 'dark';
-    },
-    hide_qty_decimals() {
-      try {
-        const saved = localStorage.getItem('posawesome_item_selector_settings');
-        if (saved) {
-          const opts = JSON.parse(saved);
-          return !!opts.hide_qty_decimals;
-        }
-      } catch (e) {
-        console.error('Failed to load item selector settings:', e);
-      }
-      return false;
-    },
+
   },
   methods: {
     onDragOverFromSelector(event) {
