+++ posawesome/public/js/posapp/components/pos/NewAddress.vue
@@ -52,18 +52,15 @@
 </template>
 
 <script>
+import { themeSettingsMixin } from '../../mixins/themeSettings.js';
 
 export default {
+  mixins: [themeSettingsMixin],
   data: () => ({
     addressDialog: false,
     address: {},
     customer: '',
   }),
-  computed: {
-    isDarkTheme() {
-      return this.$theme.current === 'dark';
-    }
-  },
 
   methods: {
     close_dialog() {
