+++ posawesome/public/js/posapp/components/pos/Payments.vue
@@ -398,10 +398,11 @@ import {
 
 import generateOfflineInvoiceHTML from "../../../offline_print_template";
 import { silentPrint } from "../../plugins/print.js";
+import { themeSettingsMixin } from '../../mixins/themeSettings.js';
 
 export default {
   // Using format mixin for shared formatting methods
-  mixins: [format],
+  mixins: [format, themeSettingsMixin],
   data() {
     return {
       loading: false, // UI loading state
@@ -570,9 +571,6 @@ export default {
       return this.pos_settings?.invoice_fields?.some(
         (el) => el.fieldtype === "Button" && el.fieldname === "request_for_payment"
       ) || false;
-    },
-    isDarkTheme() {
-      return this.$theme.current === 'dark';
     }
   },
   watch: {
