+++ posawesome/public/js/posapp/components/pos/PostingDateRow.vue
@@ -46,7 +46,9 @@
 </template>
 
 <script>
+import { themeSettingsMixin } from '../../mixins/themeSettings.js';
 export default {
+  mixins: [themeSettingsMixin],
   props: {
     pos_profile: Object,
     posting_date_display: String,
@@ -60,11 +62,6 @@ export default {
       internal_posting_date_display: this.posting_date_display,
       internal_price_list: this.priceList,
     };
-  },
-  computed: {
-    isDarkTheme() {
-      return this.$theme?.current === 'dark';
-    }
   },
   watch: {
     posting_date_display(val) {
