+++ posawesome/public/js/posapp/components/pos/Returns.vue
@@ -138,9 +138,10 @@
 
 <script>
 import format from '../../format';
+import { themeSettingsMixin } from '../../mixins/themeSettings.js';
 
 export default {
-  mixins: [format],
+  mixins: [format, themeSettingsMixin],
   data: () => ({
     invoicesDialog: false,
     singleSelect: true,
@@ -191,11 +192,6 @@ export default {
       },
     ],
   }),
-  computed: {
-    isDarkTheme() {
-      return this.$theme?.current === 'dark';
-    }
-  },
   watch: {
     from_date() {
       this.formatFromDate();
